Replace OpenSSL token generator with bin2hex + random_bytes PHP native functions
