+++ composer.json
@@ -10,7 +10,6 @@
     ],
     "require": {
         "php": "^8.1",
-        "ext-openssl": "*",
         "symfony/framework-bundle": "^6.4|^7.4|^8.0",
         "doctrine/common": "^3.0",
         "doctrine/orm": "^3.0",
