+++ config/services.xml
@@ -26,8 +26,8 @@
             <argument type="service" id="yokai_security_token.token_entity_repository"/>
         </service>
 
-        <service id="yokai_security_token.open_ssl_token_generator"
-                 class="Yokai\SecurityTokenBundle\Generator\OpenSslTokenGenerator"
+        <service id="yokai_security_token.bin2hex_token_generator"
+                 class="Yokai\SecurityTokenBundle\Generator\Bin2HexTokenGenerator"
                  public="false"/>
 
         <service id="yokai_security_token.default_information_guesser"
